% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gdlm.R
\name{coef.gdlm}
\alias{coef.gdlm}
\title{Retrieve the fitted parameters for a gdlm}
\usage{
\method{coef}{gdlm}(object, ...)
}
\arguments{
\item{object}{the gdlm object from which to retrieve a parameters}
}
\description{
Retrieve the fitted parameters for a gdlm
}
\examples{
m <- gdlm(Sepal.Width ~ Species * Petal.Width + Petal.Length, data = iris, loss = LS_LOSS())
coef(m)

}
\author{
kholub
}
