% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{compose_regularization}
\alias{compose_regularization}
\title{Compose an existing loss function with regularization}
\usage{
compose_regularization(loss, elastic_net_parameter, lambda)
}
\arguments{
\item{loss}{the loss function to be built on}

\item{elastic_net_parameter}{the weighting of L1 regularization in the elastic net. L2 will be weighted as 1 - elastic_net_parameter}

\item{lambda}{parameter controlling how much regularization contributes to the overall loss}
}
\description{
Given an arbitrary loss function, compose it with elastic net regularization
}
\examples{
lasso_ls <- compose_regularization(LS_LOSS, 1, 1e-5)
ridge_ls <- compose_regularization(LS_LOSS, 0, 1e-5)

}
\author{
kholub
}
