% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gdlm.R
\name{formula.gdlm}
\alias{formula.gdlm}
\title{Retrieve the formula from a gdlm object}
\usage{
\method{formula}{gdlm}(object, ...)
}
\arguments{
\item{object}{the gdlm object from which to retrieve a formula}
}
\description{
Retrieve the formula from a gdlm object
}
\examples{
m <- gdlm(Sepal.Width ~ Species * Petal.Width + Petal.Length, data = iris, loss = LS_LOSS())
print(summary(m))

}
\author{
kholub
}
