% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdlm.R
\name{gdlm}
\alias{gdlm}
\title{Fit a linear model using an arbitrary loss function}
\usage{
gdlm(formula, data, loss, bootstrapped_se = TRUE, bootstrap_trials = 100,
  initial_estimators = NULL)
}
\arguments{
\item{formula}{the form of the linear model to be fit}

\item{data}{the dataset used to fit the model}

\item{loss}{a function accepting a vector of parameter estimators, a matrix of training data, and a vector of responses}

\item{bootstrapped_se}{whether to perform bootstrapping to generate standard errors and confidence intervals}

\item{initial_estimators}{the starting fit used by gradient descent. default is zeros}

\item{boostrap_trials}{how many bootstrap trials to perform}
}
\description{
Fit a linear model using gradient descent methods. Given a loss function, gdlm will minimize summed loss across observations.
Optionally generate bootstrapped standard errors & confidence intervals for estimators.
}
\note{
It is left to the user to consider the convexity of the summed loss function. For non-convex loss, gradient descent methods may not be generally appropriate.

The user can control the parallelism of bootstrapping by setting
}
\examples{
m <- gdlm(Sepal.Width ~ Species * Petal.Width + Petal.Length, data = iris, loss = LS_LOSS)

}
\author{
kholub
}
