% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{LAD_LOSS}
\alias{LAD_LOSS}
\title{Parameterize a quasi least absolute deviation loss function}
\usage{
LAD_LOSS(loss_asymmetry = 0.5)
}
\arguments{
\item{loss_asymmetry}{parameter passed to lad_loss}
}
\description{
A convinience to create a function delegating to lad_loss
}
