% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{LOGISTIC_LOSS}
\alias{LOGISTIC_LOSS}
\title{Logistic loss function}
\usage{
LOGISTIC_LOSS(estimators, X, y)
}
\arguments{
\item{estimators}{coefficient estimators}

\item{X}{the data on which to evaluate loss}

\item{the}{responses corresponding to y}
}
\description{
Logistic loss function
}
\examples{
estimators <- c(-1,1)
X <- matrix(c(1,2,1.5,2.5), nrow = 2)
y <- c(1, 0)
LOGISTIC_LOSS(estimators, X, y)

}
\author{
kholub
}
