% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{LS_LOSS}
\alias{LS_LOSS}
\title{Least squares loss function}
\usage{
LS_LOSS(estimators, X, y)
}
\arguments{
\item{estimators}{coefficient estimators}

\item{X}{the data on which to evaluate loss}

\item{the}{responses corresponding to y}
}
\description{
Least squares loss function
}
\examples{
estimators <- c(1,2)
X <- matrix(c(1,2,3,4), nrow = 2)
y <- c(6, 11)
LS_LOSS(estimators, X, y)

}
\author{
kholub
}
