% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gdlm.R
\name{plot.gdlm}
\alias{plot.gdlm}
\title{Plot a gdlm object}
\usage{
\method{plot}{gdlm}(object, ...)
}
\arguments{
\item{the}{gdlm object to be ploted}
}
\description{
Plot a gdlm object
}
\examples{
m <- gdlm(Sepal.Width ~ Species * Petal.Width + Petal.Length, data = iris, loss = LS_LOSS())
plot(m)

}
\author{
kholub
}
