% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.gdlm.R
\name{predict.gdlm}
\alias{predict.gdlm}
\title{Generate predictions from a gradient descent generated linear model}
\usage{
\method{predict}{gdlm}(object, newdata = NULL, type = c("working",
  "response"), ...)
}
\arguments{
\item{object}{model to generate predictions from}

\item{newdata}{data to predict responses for}

\item{type}{the scale on which to generate residuals}
}
\description{
Generate predictions from a gradient descent generated linear model
}
\examples{
m <- gdlm(Sepal.Width ~ Species * Petal.Width + Petal.Length, data = iris, loss = LS_LOSS)
predict(m)

}
\author{
kholub
}
