% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gdlm.R
\name{summary.gdlm}
\alias{summary.gdlm}
\title{Produce a summary for a gradient descent generated linear model}
\usage{
\method{summary}{gdlm}(object, ci_range = c(0.05, 0.95), ...)
}
\arguments{
\item{object}{model to summarize}

\item{ci_range}{a vector of the lower and upper bounds for empirical confidence intervals}
}
\description{
This summary will include model details, coefficient estimates, & bootstrapped SEs and bootstrap CIs if the gdlm was constructed with boostrapping enabled
}
\examples{
m <- gdlm(Sepal.Width ~ Species * Petal.Width + Petal.Length, data = iris, loss = LS_LOSS)
summary(m)

}
\author{
kholub
}
