% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{LAD_LOSS}
\alias{LAD_LOSS}
\title{Least absolute deviation loss}
\usage{
LAD_LOSS(estimators, X, y)
}
\arguments{
\item{estimators}{coefficient estimators}

\item{X}{the data on which to evaluate loss}

\item{the}{responses corresponding to y}
}
\description{
Note that this is not a convex loss function, and so gradient descent is not guaranteed to arrive at a global minima!
}
\examples{
estimators <- c(1,2)
X <- matrix(c(1,2,3,4), nrow = 2)
y <- c(6, 11)
LAD_LOSS(estimators, X, y)

}
\author{
kholub
}
