% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{LOGISTIC_LOSS}
\alias{LOGISTIC_LOSS}
\title{Parameterize a quasi logistic loss function}
\usage{
LOGISTIC_LOSS(loss_asymmetry = 0.5)
}
\arguments{
\item{loss_asymmetry}{parameter passed to logistic_loss}
}
\description{
A convinience to create a function delegating to logistic_loss
}
