% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{LS_LOSS}
\alias{LS_LOSS}
\title{Parameterize a quasi least squares loss function}
\usage{
LS_LOSS(loss_asymmetry = 0.5)
}
\arguments{
\item{loss_asymmetry}{parameter passed to ls_loss}
}
\description{
A convinience to create a function delegating to ls_loss
}
